[Task 1]

+++ app/src/main/java/com/example/foodapp/data/repository/CartRepository.kt
@@ -7,8 +7,8 @@ import com.example.foodapp.utils.ResultWrapper
 import kotlinx.coroutines.flow.Flow
 
 interface CartRepository {
-    fun getUserCartData(): Flow<ResultWrapper<Pair<List<Cart>, Double>>>
-    fun getCheckoutData(): Flow<ResultWrapper<Triple<List<Cart>,List<PriceItem>, Double>>>
+    fun getUserCartData(): Flow<ResultWrapper<Pair<List<Cart>, Int>>>
+    fun getCheckoutData(): Flow<ResultWrapper<Triple<List<Cart>,List<PriceItem>, Int>>>
     fun createCart(product: Menu, quantity: Int, notes: String? = null) : Flow<ResultWrapper<Boolean>>
     fun decreaseCart(item : Cart) : Flow<ResultWrapper<Boolean>>
     fun increaseCart(item : Cart) : Flow<ResultWrapper<Boolean>>
